feat: refactor where helper methods are
